Incorporate differential learning rates for both roberta & custom head
Allows for "Freeze Roberta" parameter in config.yaml file to be set to false, allowing roberta to learn from training data - the learning rate for the roberta model is set to 10% of the custom head rate. Not perfect, but much better than the higher rate for the custom head. Ideally this is fine tuned during hyperparameter tuning, but that would take weeks considering intital optimization took days.